# contracts/openapi/openapi-admin.yml
openapi: 3.1.0
info:
  title: Memory Service Admin API
  version: v1
  description: |-
    Admin APIs for Memory Service. These endpoints provide administrative access
    to all conversations across all users, including archived resources.
    Access requires admin or auditor role.
...
paths:
...
  /v1/admin/conversations/{id}:
    get:
      tags: [Admin]
      summary: Get any conversation (admin/auditor)
      description: |-
        Retrieves a conversation by ID, including archived conversations.
        Requires auditor or admin role.
      operationId: adminGetConversation
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation identifier (UUID format).
          schema:
            type: string
            format: uuid
        - name: justification
          in: query
          required: false
          description: Reason for this admin action (for audit log).
          schema:
            type: string
      responses:
        '200':
          description: The conversation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConversation'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []
    patch:
      tags: [Admin]
      summary: Update any conversation (admin only)
      description: |-
        Updates admin-managed conversation state. Requires admin role.
        Setting `archived` to `true` archives the whole fork tree. Setting it to
        `false` unarchives the fork tree.
      operationId: adminUpdateConversation
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation identifier (UUID format).
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateConversationRequest'
      responses:
        '200':
          description: Conversation updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConversation'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []
  /v1/admin/conversations/{id}/entries:
    get:
      tags: [Admin]
      summary: Get entries from any conversation (admin/auditor)
      description: |-
        Retrieves entries from a conversation, including archived
        conversations. Requires auditor or admin role.
      operationId: adminGetEntries
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation identifier (UUID format).
          schema:
            type: string
            format: uuid
        - name: afterCursor
          in: query
          required: false
          description: Cursor for pagination (UUID format).
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          description: Maximum number of entries to return.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
        - name: channel
          in: query
          required: false
          description: Filter by entry channel.
          schema:
            $ref: '#/components/schemas/Channel'
        - name: justification
          in: query
          required: false
          description: Reason for this admin action (for audit log).
          schema:
            type: string
        - name: forks
          in: query
          required: false
          description: |-
            Controls which fork entries to include. `none` (default) follows
            the fork ancestry path. `all` returns entries from all forks.
          schema:
            type: string
            enum: [none, all]
            default: none
      responses:
        '200':
          description: Entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entry'
                  afterCursor:
                    type: string
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []
  /v1/admin/conversations/{id}/memberships:
    get:
      tags: [Admin]
      summary: Get memberships for any conversation (admin/auditor)
      description: |-
        Retrieves memberships for a conversation. Requires auditor or admin role.
      operationId: adminGetMemberships
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation identifier (UUID format).
          schema:
            type: string
            format: uuid
        - name: afterCursor
          in: query
          required: false
          description: Cursor for pagination; returns items after this user id.
          schema:
            type: string
            nullable: true
        - name: limit
          in: query
          required: false
          description: Maximum number of memberships to return.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
        - name: justification
          in: query
          required: false
          description: Reason for this admin action (for audit log).
          schema:
            type: string
      responses:
        '200':
          description: Memberships.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationMembership'
                  afterCursor:
                    type: string
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []
  /v1/admin/conversations/{id}/forks:
    get:
      tags: [Admin]
      summary: List forks for any conversation (admin/auditor)
      description: |-
        Returns all forked conversations that share the same root conversation as the
        given conversation. Each fork entry includes the message id at which it forked
        and the timestamp when the forked conversation was created.

        This endpoint provides admin access to fork information for any conversation,
        regardless of ownership. Requires auditor or admin role.
      operationId: adminListForks
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation identifier (UUID format).
          schema:
            type: string
            format: uuid
        - name: afterCursor
          in: query
          required: false
          description: Cursor for pagination; returns items after this conversation id (UUID format).
          schema:
            type: string
            format: uuid
            nullable: true
        - name: limit
          in: query
          required: false
          description: Maximum number of forks to return.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
        - name: justification
          in: query
          required: false
          description: Reason for this admin action (for audit log).
          schema:
            type: string
      responses:
        '200':
          description: Forked conversations related to this conversation's root.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationForkSummary'
                  afterCursor:
                    type: string
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []
  /v1/admin/conversations/{id}/children:
    get:
      tags: [Admin]
      summary: List direct child conversations
      operationId: adminListChildConversations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: afterCursor
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: A list of direct child conversations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminChildConversationSummary'
                  afterCursor:
                    type: string
                    nullable: true
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []
  /v1/admin/conversations/search:
    post:
      tags: [Admin]
      summary: System-wide semantic search (admin/auditor)
      description: |-
        Performs semantic search across all conversations, optionally filtered
        by userId. Requires auditor or admin role.
      operationId: adminSearchConversations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSearchEntriesRequest'
      parameters:
        - name: justification
          in: query
          required: false
          description: Reason for this admin action (for audit log).
          schema:
            type: string
      responses:
        '200':
          description: Search results.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResult'
                  afterCursor:
                    type: string
                    nullable: true
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    AccessLevel:
      type: string
      description: Access level of a user for a conversation.
      enum:
        - owner
        - manager
        - writer
        - reader
    AdminConversationSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the conversation.
        title:
          type: string
          nullable: true
        ownerUserId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archived:
          type: boolean
          description: Synthetic archive flag derived from the internal archived timestamp.
        lastMessagePreview:
          type: string
          nullable: true
        accessLevel:
          $ref: '#/components/schemas/AccessLevel'
        clientId:
          type: string
          nullable: true
        agentId:
          type: string
          nullable: true
        startedByConversationId:
          type: string
          format: uuid
          nullable: true
        startedByEntryId:
          type: string
          format: uuid
          nullable: true
    AdminChildConversationSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the child conversation.
        title:
          type: string
          nullable: true
        ownerUserId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archived:
          type: boolean
          description: Synthetic archive flag derived from the internal archived timestamp.
        lastMessagePreview:
          type: string
          nullable: true
        accessLevel:
          $ref: '#/components/schemas/AccessLevel'
        startedByEntryId:
          type: string
          format: uuid
          nullable: true
    AdminConversation:
      allOf:
        - $ref: '#/components/schemas/AdminConversationSummary'
        - type: object
          properties:
            forkedAtEntryId:
              type: string
              format: uuid
              nullable: true
              description: >-
                First parent entry excluded by this fork. Null for root
                conversations and blank-slate forks that inherit no parent entries.
            forkedAtConversationId:
              type: string
              format: uuid
              nullable: true
              description: Conversation ID from which this conversation was forked.
            startedByConversationId:
              type: string
              format: uuid
              nullable: true
            startedByEntryId:
              type: string
              format: uuid
              nullable: true
    AdminActionRequest:
      type: object
      properties:
        justification:
          type: string
          description: Reason for the admin action (for audit log).
      example:
        justification: "User requested account cleanup per support ticket #1234"
    AdminUpdateConversationRequest:
      allOf:
        - $ref: '#/components/schemas/AdminActionRequest'
        - type: object
          properties:
            archived:
              type: boolean
              description: Set to `true` to archive or `false` to unarchive the conversation fork tree.
    Channel:
      type: string
      description: Logical channel of the entry within the conversation.
      enum:
        - history
        - context
    Entry:
      type: object
      required:
        - id
        - conversationId
        - channel
        - contentType
        - content
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the entry.
        conversationId:
          type: string
          format: uuid
          description: Unique identifier for the conversation this entry belongs to.
        userId:
          type: string
          nullable: true
        channel:
          $ref: '#/components/schemas/Channel'
        epoch:
          type: integer
          format: int64
          nullable: true
        contentType:
          type: string
          description: |-
            Describes the schema/format of the content array.
            Examples: "message", "LC4J", "SpringAI"
        content:
          type: array
          items: {}
        createdAt:
          type: string
          format: date-time
    ConversationMembership:
      type: object
      properties:
        conversationId:
          type: string
          format: uuid
          description: Unique identifier for the conversation.
        userId:
          type: string
        accessLevel:
          $ref: '#/components/schemas/AccessLevel'
        createdAt:
          type: string
          format: date-time
    ConversationForkSummary:
      type: object
      description: Summary of a forked conversation originating at a given entry.
      properties:
        conversationId:
          type: string
          format: uuid
          description: Unique identifier for the forked conversation.
        forkedAtEntryId:
          type: string
          format: uuid
          nullable: true
          description: >-
            First parent entry excluded by this fork. Null for blank-slate forks
            that inherit no parent entries.
        forkedAtConversationId:
          type: string
          format: uuid
          nullable: true
          description: Conversation ID where the fork occurred.
        title:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    AdminSearchEntriesRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 1000
          description: Natural language query.
        afterCursor:
          type: string
          nullable: true
          maxLength: 100
          description: Cursor for pagination; returns items after this result.
        limit:
          type: integer
          default: 20
          minimum: 1
          maximum: 1000
          description: Maximum number of results to return.
        userId:
          type: string
          nullable: true
          maxLength: 255
          description: Filter search to conversations owned by this user.
        includeArchived:
          type: boolean
          default: false
          description: Include entries from archived conversations.
        includeEntry:
          type: boolean
          default: true
          description: Whether to include the full entry in results. Set to false to reduce response size when only metadata is needed.
    SearchResult:
      type: object
      properties:
        conversationId:
          type: string
          format: uuid
          description: Unique identifier of the conversation containing this entry.
        conversationTitle:
          type: string
          description: Title of the conversation containing this entry.
        entryId:
          type: string
          format: uuid
          description: ID of the matched entry. Always present for deep-linking.
        score:
          type: number
          format: float
        highlights:
          type: string
          nullable: true
        entry:
          allOf:
            - $ref: '#/components/schemas/Entry'
          description: The matched entry. Only included when includeEntry is true in the request.
...