# java/quarkus/examples/chat-quarkus/src/main/java/org/acme/ConversationsResource.java
package org.acme;

import io.github.chirino.memory.client.model.Channel;
import io.github.chirino.memory.runtime.MemoryServiceProxy;
import io.smallrye.common.annotation.Blocking;
import jakarta.enterprise.context.ApplicationScoped;
import jakarta.inject.Inject;
import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.DELETE;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.PATCH;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;

/**
 * JAX-RS resource that proxies requests to the /v1/conversations endpoint so that the frontend can
 * access the conversation history.
 */
@Path("/v1/conversations")
@ApplicationScoped
@Blocking // Offload REST client calls from the event loop to prevent deadlock
public class ConversationsResource {

    @Inject MemoryServiceProxy proxy;

    @GET
    @Produces(MediaType.APPLICATION_JSON)
    public Response listConversations(
            @QueryParam("mode") String mode,
            @QueryParam("ancestry") String ancestry,
            @QueryParam("afterCursor") String afterCursor,
            @QueryParam("limit") Integer limit,
            @QueryParam("query") String query,
            @QueryParam("archived") String archived) {
        return proxy.listConversations(mode, ancestry, afterCursor, limit, query, archived);
    }

    @GET
    @Path("/{conversationId}")
    @Produces(MediaType.APPLICATION_JSON)
    public Response getConversation(@PathParam("conversationId") String conversationId) {
        return proxy.getConversation(conversationId);
    }

    @PATCH
    @Path("/{conversationId}")
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.APPLICATION_JSON)
    public Response updateConversation(
            @PathParam("conversationId") String conversationId, String body) {
        return proxy.updateConversation(conversationId, body);
    }

    @GET
    @Path("/{conversationId}/entries")
    @Produces(MediaType.APPLICATION_JSON)
    public Response listConversationEntries(
            @PathParam("conversationId") String conversationId,
            @QueryParam("afterCursor") String afterCursor,
            @QueryParam("limit") Integer limit) {
        return proxy.listConversationEntries(
                conversationId, afterCursor, limit, Channel.HISTORY, null, "all");
    }

    @GET
    @Path("/{conversationId}/forks")
    @Produces(MediaType.APPLICATION_JSON)
    public Response listConversationForks(@PathParam("conversationId") String conversationId) {
        return proxy.listConversationForks(conversationId, null, null);
    }

    @GET
    @Path("/{conversationId}/children")
    @Produces(MediaType.APPLICATION_JSON)
    public Response listConversationChildren(
            @PathParam("conversationId") String conversationId,
            @QueryParam("afterCursor") String afterCursor,
            @QueryParam("limit") Integer limit) {
        return proxy.listConversationChildren(conversationId, afterCursor, limit);
    }

    @GET
    @Path("/{conversationId}/memberships")
    @Produces(MediaType.APPLICATION_JSON)
    public Response listConversationMemberships(
            @PathParam("conversationId") String conversationId) {
        return proxy.listConversationMemberships(conversationId, null, null);
    }

    @POST
    @Path("/{conversationId}/memberships")
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.APPLICATION_JSON)
    public Response shareConversation(
            @PathParam("conversationId") String conversationId, String body) {
        return proxy.shareConversation(conversationId, body);
    }

    @PATCH
    @Path("/{conversationId}/memberships/{userId}")
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.APPLICATION_JSON)
    public Response updateConversationMembership(
            @PathParam("conversationId") String conversationId,
            @PathParam("userId") String userId,
            String body) {
        return proxy.updateConversationMembership(conversationId, userId, body);
    }

    @DELETE
    @Path("/{conversationId}/memberships/{userId}")
    public Response deleteConversationMembership(
            @PathParam("conversationId") String conversationId,
            @PathParam("userId") String userId) {
        return proxy.deleteConversationMembership(conversationId, userId);
    }

    @POST
    @Path("/search")
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.APPLICATION_JSON)
    public Response searchConversations(String body) {
        return proxy.searchConversations(body);
    }
}